Refactor: Introduce SteamStoreService for store-related API interactions in SteamTools
